+++ WeChatBot.Net/Model/API/Base/BaseRequest.cs
@@ -2,7 +2,7 @@ namespace WeChatBot.Net.Model.API.Base
 {
     public class BaseRequest
     {
-        public int Uin { get; set; }
+        public long Uin { get; set; }
         public string Sid { get; set; }
         public string Skey { get; set; }
         public string DeviceID { get; set; }
